+++ nix/modules/default.nix
@@ -1,0 +1,7 @@
+{ ... }:
+{
+  imports = [
+    ./api.nix
+    ./backends/actual.nix
+  ];
+}